backend/requirements.txt
django == 5.2.3
djangorestframework == 3.16.0
requests == 2.32.4
django-cors-headers == 4.7.0
mysqlclient == 2.2.7
djangorestframework-simplejwt == 5.5.0
python-dotenv == 1.0.0
